+++ lib/vm/src/vmcontext.rs
@@ -340,7 +340,7 @@ pub struct VMMemoryDefinition {
     pub base: *mut u8,
 
     /// The current logical size of this linear memory in bytes.
-    pub current_length: u32,
+    pub current_length: u64,
 }
 
 /// # Safety
@@ -379,10 +379,10 @@ impl VMMemoryDefinition {
         // https://webassembly.github.io/reference-types/core/exec/instructions.html#exec-memory-copy
         if src
             .checked_add(len)
-            .map_or(true, |n| n > self.current_length)
+            .map_or(true, |n| u64::from(n) > self.current_length)
             || dst
                 .checked_add(len)
-                .map_or(true, |m| m > self.current_length)
+                .map_or(true, |m| u64::from(m) > self.current_length)
         {
             return Err(Trap::new_from_runtime(TrapCode::HeapAccessOutOfBounds));
         }
@@ -412,7 +412,7 @@ impl VMMemoryDefinition {
     pub(crate) unsafe fn memory_fill(&self, dst: u32, val: u32, len: u32) -> Result<(), Trap> {
         if dst
             .checked_add(len)
-            .map_or(true, |m| m > self.current_length)
+            .map_or(true, |m| u64::from(m) > self.current_length)
         {
             return Err(Trap::new_from_runtime(TrapCode::HeapAccessOutOfBounds));
         }
